requirements.txt
annotated-types==0.7.0
blinker==1.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.28.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
oauthlib==3.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic_core==2.27.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
PyYAML==5.4.1
requests==2.32.3
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
tqdm==4.67.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
